WebSite Update
The head is the information that is tells the browser whats in the webSite. There is a meta that has keywords that helps the browser commit a faster search for the website.The bod has a header inside and in the header there is a navigation list with an unordered list of index, Info, FAQ and these help go to different areas and the items have certain names that are connected to the items.The header has two sections that have information about how much money gets a person what they need. The title is in the h1 tag which is the primary importance. The amount of money that the care is, is in the strong tag which will make the amount of money bold.The second section has a header, primary importance, secondary importance, img element, Strong tag, paragraph tags. The header has the title of the information and and it is the primary importance text. The image element is telling the computer that the image is supposed to be there and so it is linking the image to the page. The secondary importance tags have the the text that is supposed to be bold. This is just telling the system that the text is supposed to be bold but no the most bold. The paragraph tags just keeps the text separated from the other text.The footer has the contact information and has the character code. The character code breaks the text up but not just space, but by simple symbols.The section is not fully finish but it has a piece of the text in the section. The date is in a time element with a datetime attribute with the text Tuesday, August 29, 2017