WebSite Updates.
The date is in a time element with a datetime attribute with the text Tuesday, August 29, 2017. The text "race information" is in a primary importance tag. There is a inline image with the dr_logo.png that is connected to the page with an alternative text of Diane's Run.The text below is the information about the program. The paragraphs are in paragraph tags to keep them in proper form. The titles of the paragraphs are in secondary importance tags. after the text are in paragraph tags, the information is put in a unordered listThe footer has the contact information and has the character code. The character code breaks the text up but not just space, but by simple symbolsAdded the links and the Hrefs connecting them to dr_base.css and dr_layout3.css.Added the header into the body from the index page so same header.Under the header there is a ordered list with the thirteen questions. The thirteen questions are in hyperlinks and the hrefs are the faq and the number of the question. Then all of the thirteen questions are in listed item tags. Then under those are the thirteen questions with the paragraphs that are in paragraph tags. The heading are in secondary importance tags. But inside of the of the tag there is an id attribute. This sets the particular section set to the certain faq.